move radius functionality to new util.py module
