+++ legacyhalos/util.py
@@ -1,0 +1,39 @@
+"""
+legacyhalos.util
+================
+
+Miscellaneous utility code used by various scripts.
+
+"""
+from __future__ import absolute_import, division, print_function
+
+import numpy as np
+
+def cutout_radius_100kpc(redshift, pixscale=0.262, radius_kpc=100):
+    """Get a cutout radius of 100 kpc [in pixels] at the redshift of the cluster.
+
+    """
+    from astropy.cosmology import WMAP9 as cosmo
+    arcsec_per_kpc = cosmo.arcsec_per_kpc_proper(redshift).value
+    radius = np.rint(radius_kpc * arcsec_per_kpc / pixscale).astype(int) # [pixels]
+    return radius
+
+def cutout_radius_cluster(redshift, cluster_radius, pixscale=0.262, factor=1.0,
+                          rmin=50, rmax=500, bound=False):
+    """Get a cutout radius which depends on the richness radius (in h^-1 Mpc)
+    R_LAMBDA of each cluster (times an optional fudge factor).
+
+    Optionally bound the radius to (rmin, rmax).
+
+    """
+    from astropy.cosmology import WMAP9 as cosmo
+
+    radius_kpc = cluster_radius * 1e3 * cosmo.h # cluster radius in kpc
+    radius = np.rint(factor * radius_kpc * cosmo.arcsec_per_kpc_proper(redshift).value / pixscale)
+
+    if bound:
+        radius[radius < rmin] = rmin
+        radius[radius > rmax] = rmax
+
+    return radius
+
